Rename worker-related models
