+++ src/Model/Machine/State.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Model\Worker;
+namespace App\Model\Machine;
 
 class State implements \Stringable
 {
