+++ src/Services/MachineHandler/CreateMachineHandler.php
@@ -8,10 +8,10 @@ use App\Entity\Machine;
 use App\Message\MachineRequestMessage;
 use App\MessageDispatcher\MachineRequestMessageDispatcherInterface;
 use App\Model\ApiRequestOutcome;
+use App\Model\Machine\State;
 use App\Model\MachineProviderActionInterface;
 use App\Model\MachineRequest;
 use App\Model\MachineRequestInterface;
-use App\Model\Worker\State;
 use App\Repository\MachineRepository;
 use App\Services\ApiActionRetryDecider;
 use App\Services\ExceptionLogger;
