+++ tests/Functional/Controller/WorkerControllerTest.php
@@ -7,9 +7,9 @@ namespace App\Tests\Functional\Controller;
 use App\Controller\WorkerController;
 use App\Entity\Machine;
 use App\Message\MachineRequestMessage;
+use App\Model\Machine\State;
 use App\Model\MachineRequest;
 use App\Model\ProviderInterface;
-use App\Model\Worker\State;
 use App\Repository\MachineRepository;
 use App\Request\MachineCreateRequest;
 use App\Services\MachineFactory;
